ccsr/pyproject.toml
[project]
name = "ccsr"
version = "0.1.0"
description = "Contrastive Curriculum Sentiment Reasoning for Healthcare"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "datasets>=2.19.0",
    "accelerate>=0.29.0",
    "deepspeed>=0.14.0",
    "evaluate>=0.4.0",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    "pyyaml>=6.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",
    "jupyter>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=5.0.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
